db update

+++ climbing/climbing/src/main/java/fr/climbing/model/Role.java
@@ -6,17 +6,19 @@ import java.util.Set;
 @Entity
 @Table(name = "role")
 public class Role {
-    private Long id;
+    private int id;
     private String name;
     private Set<User> users;
 
+  
+    
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
-    public Long getId() {
+    public int getId() {
         return id;
     }
 
-    public void setId(Long id) {
+    public void setId(int id) {
         this.id = id;
     }
 
@@ -36,4 +38,6 @@ public class Role {
     public void setUsers(Set<User> users) {
         this.users = users;
     }
+    
+ 
 }
